fix handling of descriptors with __set__ but no __get__
Summary:
Python divides descriptors into two kinds: "data descriptors" and "non-data descriptors." Data
descriptors take precedence over entries in the instance dict, non-data descriptors do not.

The [Python descriptor documentation](https://docs.python.org/3/howto/descriptor.html) says
"If an object defines `__set__()` or `__delete__()`, it is considered a data descriptor," and there is a
`PyDescr_IsData()` macro that checks exactly this definition.

But there is a subtlety here: it is possible to create a descriptor that defines `__set__` (and/or
`__delete__`) but does NOT define `__get__`. Although this is technically a "data descriptor" by
the documented definition, in practice it isn't: since it doesn't define `__get__`, it should play no
role in `LOAD_ATTR` and it should not override a value present in the instance dict.

Shadowcode and jit inline caches were both getting this case wrong,  treating any
descriptor with a `tp_descr_set` as a data descriptor. In the JIT, this case would simply crash due
to `DataDescrMutator` blindly assuming the presence of a `tp_descr_get`. Shadowcode guarded
against null `tp_descr_get` but still wrongly overrode the value from the instance dict, and if
the descriptor lacked `__get__`, it would fall back to returning the descriptor itself.

The JIT fix is to ensure we fall back to a regular `DescrOrClassVarMutator` (instead of
`DataDescrMutator`) in the set-but-no-get case; this mutator implements full correct lookup logic
and doesn't assume a `tp_descr_get`.

In shadowcode, ensure that when `__get__` is missing, a value from the instance dict (if any) is
returned, and we only return the descriptor object itself if there was no value in the instance dict.

This fix needs to be made in two separate cases in shadowcode: split and combined dictionaries.

Also made the structure of `_PyShadow_LoadAttrDictDescrHit` a bit more similar to
`_PyShadow_LoadAttrSplitDictDescrHit`.

D44279728 is the 3.8 version of this diff.

Reviewed By: swtaarrs

Differential Revision: D44281234

fbshipit-source-id: 7e67ba36a31b0186e3292bdd26f2979ef2f56c01