Hydrate lazy objects
Summary:
When the IGSRV server starts, there are around 840K lazy objects across all loaded modules. After preloading, ~93% of them can be hydrated so workers don't have to update the dictionaries of the modules the first time a lazy object is accessed.

This should reduce the number of CoWs we see originating from accesses to lazy imported objects in the workers. The drawback is that hydrating these objects in master adds around 7 seconds.

 ---

In this implementation, `importlib.hydrate_lazy_objects()` tries to hydrate the lazy objects associated with already loaded modules.

Reviewed By: DinoV

Differential Revision: D44220896

fbshipit-source-id: aa8e36cd1ff2fe017eb90959f0cc62cb993315c0