Fix crash when patching class method
Summary:
For handling of class methods we currently have a code path which is used after we've determined we don't really have a class method (i.e. it's been patched away).  But we still treat it like a class method, and we crash.

Now we just use a generic call against the patched value.

Reviewed By: carljm

Differential Revision: D44649197

fbshipit-source-id: 7c2ecedb4fa83b4c47d8f5385e53e9ff2a2f6d11