Generalize GetLoadMethodInstance into GetSecondOutput
Summary:
`hir::GetSecondOutput` has no restrictions on placement and can be
used to extract the second part of any call-based instruction that returns
values in two registers. It's lowered into the `LoadSecondCallResult` LIR
instruction, which has no restrictions except that there can be at most one
`LoadSecondCallResult` consuming the result of each `Call` (it's possible to
support this if needed, just not worth the complexity for now).

A new post-generation LIR rewrite function replaces `LoadSecondCallResult` with
a `Move` from the appropriate register, placed immediately after the `Call`
(which used to be done manually in `LIRGenerator`). If there is a `Phi` joining
the output of multiple `Call`s, the rewrite inserts a parallel `Phi`, repeating
as needed (this won't happen naturally until D43723041).

Reviewed By: mengdilin

Differential Revision: D44594084

fbshipit-source-id: 81140ebc30a7b1f6cdc0cdb461e9bb65c23df185