Don't do LOAD_TYPE in generated code for class methods
Summary: Currently when generating a call to a `classmethod` we emit into the generated call a `LOAD_TYPE` on `self` if we're doing the call with an instance instead of a class. This results in us not having `self` in hand if an overload or a patch causes the target not to be a `classmethod`.  Now we check to see if we have an instance and if so then we get the class instead.

Reviewed By: carljm

Differential Revision: D44651265

fbshipit-source-id: 74ba5f9aa91bcebece1a333ad3e341bbad025e62