Start of native entry points for methods
Summary:
This is the first step in implementing static entry points for methods so that we can use the native entry point and by-pass the argument type checks when we're invoking methods.  This sets up the basic system we'll use, which is identical to the one used for JITed functions - that is the entrypoint has a static entrypoint that is 11 bytes earlier than the vector call entry point.

For methods, when we need a helper thunk, we setup a series of entrypoints using inline assembly which can either dispatch to a native helper or a vectorcall entry point.  With this change we still only use the vectorcall entry point, and there's no updates to JITed code to change how we call into these.  That'll come in the next diff.

Reviewed By: carljm

Differential Revision: D43969361

fbshipit-source-id: e083bc63033f4386f8b3512951a2257dca84e478