Add static method invoke
Summary: This adds the support for the native entry points in the method v-tables.  We have various helpers that will need to re-hydrate the arguments and dispatch into the function using the vectorcall version.  The most common scenario would be if we're not JITed the targetted functions.  But things like classmethods, staticfunctions, non-sealed types, objects with dictionaries, etc... will all hit the same limitations.  If any of these become perf hot spots we can get more custom helpers which do whatever adjustments are needed to the incoming registers.

Reviewed By: carljm

Differential Revision: D44231824

fbshipit-source-id: 5cd9c031478ddc02c186061cb52f6d7b08478fa1