Add _ALVResultFuture type to represent result of in-flight AsyncLazyValue
Summary:
This is a first diff in the stack that enables cycle detection between running `AsyncLazyValue`. The gist of the idea:
if code attempts to await the result of in-flight `AsyncLazyValue` #ALV0 - we need to check if we are currently running as a part of some another `AsyncLazyValue` #ALV1. If yes - we need to validate that execution of #ALV0 is not transitively blocked on #ALV1 - otherwise execution would never end.

`AsyncLazyValue` uses futures if somebody it attempting to await on running `AsyncLazyValue`. We would need to augment returned future with some additional information (i.e. for keeping track what `AsyncLazyValue` is blocked on this future). Rather than extending standard `Future` type this diff would add a dedicated `_ALVResultFuture` type which behaves pretty much exactly as `Future` (sharing most of the implementation with it) - my rationale is this would make upgrade process easier since it is simpler to just copy type declaration rather than doing fine-grained tracking of differences in implementation of functions.

Reviewed By: mpage

Differential Revision: D45173446

fbshipit-source-id: 202cbd4dc95f20d8a66e8acbec1271f74e74cafd