Split _imp._maybe_set_parent_attribute function for adding parent and child side effect separately
Summary:
Lazy Imports side effects are added by `_find_and_load_unlocked()` after the module has been imported. We used to have a single function to add side effects to the parent module as well as to the current module. This diff splits `_imp._maybe_set_submodule_attribute()` into two functions:

1. `_imp._maybe_set_parent_attribute()`, for maybe adding an attribute in the parent module, pointing to the imported module (if it's still needed there at the time of the lazy object resolution).
2. `_imp._set_lazy_attributes()`, for adding attributes in the imported module, pointing to any lazy child modules.

This is to have clearer warnings when any of the two fails, and to split different (albeit related) functionality to different functions.

Differential Revision: D45825784

fbshipit-source-id: 151fd542585a7bffaa62bb316257f14c78359d6e