User error codes and fix 32-bit system issues
Summary:
Port the latest changes from the [perf-map C-API PR](https://github.com/python/cpython/pull/103546/). This includes using error codes instead of exceptions to avoid the need to hold the GIL.

Additionally, there was an issue with Fedora on a 32-bit system that was fixed in this [PR](https://github.com/python/cpython/pull/104811), so I also ported the fix.

Reviewed By: carljm

Differential Revision: D46041553

fbshipit-source-id: 6502189f489ca6cf11949ecd93c044e5ebe4fd2c