Cycle detection when AsyncLazyValue is immediately awaited
Summary:
This diff adds initial implementation of cycle checking when running `AsyncLazyValue` is awaited directly or via result of `ensure_future` (support for doing this in `gather` is coming in subsequent diff)

Idea of the implementation:
We augment `AsyncLazyValue` so when it is started we try to find `AsyncLazyValue` that is currently running - it would be recorded in `AsyncLazyValue.parent` attribute. Lookup is done via stack walk - we use dedicated Python entrypoint that will not be JIT-ted so we can find Python frame for it during the stack walk and pull local value from a known slot.

When `ALVResultFuture` #1 is awaited we again look for `AsyncLazyValue` #2 that is currently running. If it is not found - this means that code is executed outside of `AsyncLazyValue` and we can continue as-is. Otherwise we set the bit on future #1 so it could be distinguished later and do a recursive walk over futures in `AsyncLazyValue`'s starting from #2 trying to find #1. If we succeed - this means that `AsyncLazyValue` that supposed to provide value for #1 is already blocked expecting result from #2 so computation will never finish.

Reviewed By: mpage, carljm

Differential Revision: D45214461

fbshipit-source-id: 82bb92f28305ea559a5283a81ad4a9304395afaf