fix super() calls on unusual types (e.g. meta-types)
Summary:
This bug was discovered in the upstream `LOAD_SUPER_ATTR` implementation
in 3.12b1, and it also exists in the Cinder implementation.

`_PyObject_GetMethod` checks whether the type has a non-standard
`tp_getattro` (i.e. not `PyObject_GenericGetAttr`), and if so, refuses to attempt
the load-method optimization. We didn't have this check in the `super()`
optimization, causing incompatible behavior in the case of a `super()` lookup
on a type with unusual `tp_getattro`, e.g. meta-types.

Upstream fix is https://github.com/python/cpython/pull/105094

Reviewed By: mpage

Differential Revision: D46281340

fbshipit-source-id: 41aba762d3e2925042a3d897a690317ba2767cf2