add func event for qualname modification
Summary:
When the `__qualname__` of a function is modified (this can happen e.g. with a decorator using `functools.wraps`), we re-initialize the function entrypoint so we can reconsider (based on the new qualname) whether the function qualifies for jit compilation.

We missed the need for this when we upstreamed function watchers, so we'll have to upstream this additional event.

In this case we really do need to call `PyEntry_init` with the new information, so I just assign the new qualname "early" (even though `func_set_qualname` will reassign it again).

Reviewed By: alexmalyshev

Differential Revision: D47164762

fbshipit-source-id: 631b64a343b03744544a1f3a921d16fffa6d6a01