Fix deadlock in concurrent.futures:ProcessPoolExecutor
Summary:
concurrent.futures:ProcessPoolExecutor can deadlock if its call queue's
thread is blocked writing to the underlying pipe and the process pool
is marked as broken.

Some background is required in order to understand how we end up in
this state:

- concurrent.futures.process:ProcessPoolExecutor creates a _SafeQueue to
  feed work items to worker processes in the pool.
- When an item is enqueued into the _SafeQueue it is first placed into a
  deque and a worker thread (internal to the _SafeQueue) is notified that
  there are new items in the queue.
- The _SafeQueue's worker thread transfers items from the deque to worker
  processes by pickling them and writing them to the write end of a pipe (on unixes).
- Worker processes read serialized work items from the read end of the pipe.
- The ProcessPoolExecutor spawns an _ExecutorManagerThread to transfer
  items between the pool and the _SafeQueue.

The _SafeQueue's worker thread can become permanently blocked if it
tries to write to a pipe whose buffer is full and all worker processes
are dead. The latter half of this causes the pool to enter a broken
state. This can manifest as a deadlock in one of two ways:

1. Someone calls `shutdown(wait=True)` on the pool. This will join the
   pool's _ExecutorManagerThread, which is blocked waiting for its
   _SafeQueue's worker thread to finish, which is blocked writing to
   the pipe.
2. The process exits. Internally, the concurrent.futures.pool module
   registers functions that run at process exit that join all
   _ExecutorManagerThreads. We end up in (1) when this happens.

We avoid the deadlock by closing the read end of the queue's pipe when
the pool enters a broken state. Writes to the pipe will fail with EPIPE
instead of blocking once all processes have died.

Reviewed By: itamaro

Differential Revision: D47311841

fbshipit-source-id: cca75c8b19ead920a09470313f762dfa8e422d96