Make the size of objects within Slab configurable
Summary:
Currently objects inside of Slab<T> are always sized as

  roundUp(sizeof(T), alignof(T))

so the size of each slab-allocated object must be known at compile-time.

However if we have an object whose size gets determined at after the process has
started running (e.g. computed based on argv), then we can't use it with
SlabArena as it is today.

This diff makes the logic for allocating slab objects more general by offloading
that responsibility onto a trait type, so we can control it at runtime.

Reviewed By: swtaarrs

Differential Revision: D46867207

fbshipit-source-id: b129972372e37bfb012e7ad5ac34bcf3353f1258