Break out GlobalCache into its own set of files
Summary:
There's a dependency chain from pyjit.h -> hir/preload.h -> inline_cache.h, and
it's because of the GlobalCache class.  I'm aiming to make use of pyjit.h inside
of inline_cache.h in a later diff, so I need to split things up.

GlobalCache is different enough from the other definitions in inline_cache.h
that it makes sense to separate it out.

Reviewed By: swtaarrs

Differential Revision: D46867219

fbshipit-source-id: 1298b8580f514b27b53e3c2d9e2523055ae933e1