Fix Strict Modules tests not running with LSAN suppressions
Summary:
These functions are known to leak, hence why we have them in the suppressions
file.  However the strict module tests binary wasn't being configured with it.

The way this works today is that we depend on our __asan_default_options()
function in pyjit.cpp to set detect_leaks=0. With this diff we can drop that function
and fully work off of the suppressions list.

Reviewed By: swtaarrs

Differential Revision: D47973242

fbshipit-source-id: a8a553da73577ee4671f3d0f0842f39029e906a1