move Cinder/module to CinderX/
Summary:
I want to add a compilation unit under `Cinder/` for general Cinder infra that is not specifically part of the JIT, Static Python, or Strict Modules. But currently all Cinder compilation units have to be symlinked into the `Cinder/module` directory, since we have the dual build setup where we can either build the Cinder module standalone, or build Cinder features into the main runtime the way we've always done. And I don't want to introduce a cyclic symlink from `Cinder/module/Cinder -> Cinder/`. So instead move `Cinder/module/` out to a separate top-level `CinderX/` to avoid this cycle.

The naming between "Cinder" and "CinderX" is a bit confusing here, but my intended logic is that "CinderX/" is specifically the extension module, whereas "Cinder/" is general Cinder code that can be built either as part of CinderX or into the runtime (though eventually only as part of CinderX, at which point the symlinks to `Cinder/`, `Jit/`, etc inside `CinderX` will become the sole real locations of those directories.)

Reviewed By: alexmalyshev

Differential Revision: D48019256

fbshipit-source-id: a65ab67c05c76d177b543fb8885a8b497b75d50f