use code watcher in JIT
Summary:
This reverts the revert commit 94d0dbf6891cff0324fdb2d75e9d0fcc9529675f
(D47448305), re-landing D47201503, but with a key change: in the first version I'd missed the line that actually initializes the code watcher on JIT startup, so the code watcher never took effect. This meant that in a JIT-list scenario, we didn't ever remove code objects from the to-be-compiled registry when they were destroyed, so we ended up with dangling pointers on the registry.

This situation should be better tested, and I think it would have been likely caught by the multithreaded-compile test, but that test has been broken and disabled for quite a while. I created T158334137 to track fixing and re-enabling it.

This diff also adds a small improvement to the debug message on a JIT_CHECK.

I also had to fix the code watchers test suite; it didn't work correctly when another code watcher was already enabled. I upstreamed this fix in https://github.com/python/cpython/pull/107821

Reviewed By: alexmalyshev

Differential Revision: D47459869

fbshipit-source-id: c53c9847a300892d1e0b01a6a75f524ed730fd48