Support fetching type profiling info from the interpreter
Summary:
Currently we can only emit HIR based with type profiling information that's
loaded from a file.  This diff adds support for using the type profiling
information that's stored within the process's `TypeProfiler` objects.

The `HIRBuilder` will still prioritize using the type profiling information
that's been deserialized from a file.  However if that doesn't exist, it will
now check the map of `TypeProfiler` objects to see if there's a matching profile
stored there, and emit guards based on that information.

I had to remove part of the ProfileRuntime test that was verifying that
profiling information was lost after a type changed its name.  This should be
fine when using profiling information from execution time.

Reviewed By: swtaarrs

Differential Revision: D47607557

fbshipit-source-id: 41ef13e13f21e55858ef9ccf338186dc01aa391a