Move comparison operators to BCOffsetBase subclasses
GCC 13 with C++20 thinks that the operators are ambiguous as the LHS and RHS can
be reversed.  This seems to be happening because of the CRTP usage for these
classes.  Moving the operators down to the subclass level, as well as filling
out the `auto` return type, seems to get them working again.