Move comparison operators to subclasses of BCOffsetBase
Summary:
Newer compilers get unhappy with this use of comparison operators and CRTP.
They can't seem to wrap their heads around inferring the T in `BCOffsetBase<T>`
for these operators, something about how in C++20 the operators are tried twice,
once with LHS vs RHS and once with RHS vs LHS.

The easy solution is to move the operators to the concrete subclasses and call it
a day.  GCC also doesn't like the use of the `auto` return type so that's made explicit.

Reviewed By: mpage

Differential Revision: D48442851

fbshipit-source-id: 60aed0ef88d7ee1a85f14be87f7ce1bdbf682e39