Support profiling functions before sending them off to AutoJIT
Summary:
Add a new argument that tells AutoJIT to profile functions for a number of calls
before compiling them.

The `ProfileRuntime` class now tracks a set of code objects that have been
marked as good candidates for type profiling.  This is managed by the AutoJIT
handler.  The interpreter will check against this to determine which bytecodes
need to be profiled.

In order to not go through the interpreter tracing codepath for all bytecode
execution (`goto tracing_dispatch`), the interpreter will now track the number
of functions being profiled.  If the number hits 0, it will unset the
`use_tracing` bit and we will only hit the tracing path for the first bytecode
in every function rather than every bytecode.

For the previous use-case of profiling all bytecodes with a set sampling rate,
the `profile_interp` field will be set to the special value of -1.

Reviewed By: mpage

Differential Revision: D48329147

fbshipit-source-id: c338c8adbced5614a08982fb8f19cec6829673d4