Avoid setting a globals cache value to a lazy import object
Summary:
When a lazy import happens, dict watchers fire with the value set to the lazy import object itself. Thus, if a lazy import occurred after a JIT/shadowcode globals cache was already populated, we would replace the cached global value with the lazy import object itself, causing errors in subsequent code using that globals cache.

Since it is generally rare to have imports happen late in a module's lifetime, it is safe to just disable the related globals caches when this happens.

It seems like the way Bento re-runs a cell on second and subsequent executions triggers this; imports at the top of the module are re-executed but in the context of the old module; previous code objects (with their optimizations, such as shadowcode globals caches) seem to be re-used.

Reviewed By: Kronuz

Differential Revision: D48549403

fbshipit-source-id: d54cdc73ca9431520853fd044ff542ee1ff9f5df