fix sys.modules watching
Summary:
We added a `sys.modules` watcher for Static Python in D36547418 to deal with issues that can happen when an import is aborted due to an exception, but the exception is caught and the process continues, and the same module is later imported successfully.

This case is much less common in practice now than it used to be, because we fixed lazy imports module preloading to no longer silently ignore imports that fail due to a cycle. But technically it could still happen, so this `sys.modules` watcher is necessary for correctness.

But because D36547418 did not include a test, and this case doesn't occur in IG anymore, we missed porting a key line (the actual call to `_PyClassLoader_NotifyDictChange`) from 3.8 to 3.10, so the `sys.modules` watching has not actually ever been functional in 3.10.

Fix that, and add a test that fails without this line, so we don't miss this in future.

Reviewed By: mpage

Differential Revision: D48656124

fbshipit-source-id: 500ca5da5806e10bc7389f58a15f88cdccc80ee2