Build CinderX as a standalone extension module
Summary:
The current CinderX code is built and linked in as an `.so` shared library.

This change is an initial step to enable importing C/C++ code from pure-Python.

The idea is that the top-level `cinderx` module will be pure-Python and re-export values from the `_cinderx` extension module as necessary.

Reviewed By: carljm

Differential Revision: D48695374

fbshipit-source-id: c7c3fc441ec2084276d953c998784a4754509d96