re-land removal of INVOKED_STATICALLY flag
Summary:
This reverts D47297607, which re-lands D46879336. The production error that was initially blamed to this diff later resurfaced with this diff already reverted, so it must have had a different cause.

V2 -> V3 contains additional changes which are required due to changes in IG that exposed some bugs, particularly in related to overriding properties with typed descriptors with values.

Reviewed By: carljm, alexmalyshev

Differential Revision: D48286787

fbshipit-source-id: 614610f59ef9ff1b9bf58218dbc8af1e522c9baa