add support for initial jitlist to jitlist_bisect.py
Summary:
I was debugging a failure in JIT compilation. I wanted to bisect the jitlist, but `jitlist_bisect.py` didn't really support this. Because the failure happened in compilation, the jitlist autodetection never got a complete list (and crucially never included the function actually failing to compile.)

This diff adds a `--jit-list-file` option which bypasses initial jitlist detection, instead just starting with the provided initial JIT list (and verifying that it does fail.)

Reviewed By: mpage

Differential Revision: D48890325

fbshipit-source-id: 9eeada5005440ec3873590dd35423d5fa981ab77