fix awaited async call with EXTENDED_ARG
Summary: Our assertion about the shape of awaited-call bytecode did not account for the possibility of `EXTENDED_ARG`. Our bytecode iterator already does accumulate `EXTENDED_ARG`, but the `.at()` indexing method does not; rearrange the assertion so we are asserting on the result of advancing the iterator rather than on a numerical index.

Reviewed By: mpage

Differential Revision: D48892640

fbshipit-source-id: 7802b68d9e164b3f259c2e04c0df56985c162af4