Move argument checks into vector call entry point
Summary:
This moves our existing `CHECK_ARGS` checks into an entry point for non-JITed functions and sets that entry point (`Ci_StaticFunction_Vectorcall`) when we first do the initialization.  This does nothing to get rid of the opcode yet (it just becomes a nop).

There's no JIT changes necessary (other than to not use `_PyFunction_Vectorcall` directly) because the JIT already emits code for `CHECK_ARGS` as part of the prologue and not by interpreting the byte code.

Reviewed By: carljm

Differential Revision: D48878875

fbshipit-source-id: ba4ed051491aa1cc0e3d9f357851c86d7af6fc6d