Add argument type information into last entry in co_consts with return type
Summary: This adds the argument typing information into the last entry in `co_consts` replacing the existing return type descriptor with a tuple of (arg_info, return_type).  We still are emitting the existing `CHECK_ARGS` with it's redundant information.  But we now start using a `_PyClassLoader_GetCodeArgumentTypeDescrs` API which is similar to the return API to get this data.

Reviewed By: carljm

Differential Revision: D48881528

fbshipit-source-id: 03149217a4bab9c6aa389f027d0b6ebd72f20d4c