Add a function to retrieve the frame of the awaiter
Summary:
This diff adds a new function called `get_awaiter_frame` to the cinder module in CPython. The purpose of this function is to retrieve the frame of the awaiter. If there is no awaiter frame or if the frame is not a Python frame, the function returns None. Otherwise, it returns the PyFrameObject representing the awaiter frame.

The main motivation for doing this is to enhance [InstaProfiler](https://www.internalfb.com/code/fbsource/[593f582257db7ad1e49d74621b7334fab93eb907]/fbcode/instagram-server/distillery/profiling/request_profilers/insta_profiler.py?lines=172) with latency metrics, such as CPU time. To achieve this, we need to maintain a tree of the call graph during request handling and calculate the CPU time of Python functions at the end of the request. For regular functions, we use `f_back` to get the parent function. However, for calls across tasks, `f_back` will be empty, so we need to get the awaiter pointer instead.

Reviewed By: czardoz

Differential Revision: D48529888

fbshipit-source-id: 12e4e0101b65410ec5766ae80aea6ed2e0101231