stop clearing classloader cache in type_cache_clear
Summary:
We had talked about how we could do this instead by monkeypatching `sys._type_cache_clear`, but after looking at it a bit I don't think we even need to do that? There's no inherent relationship between the type method cache and the classloader cache that means the latter needs to be cleared for correctness when the former is cleared. And tests seem fine with just not doing it. We already expose `cinder.clear_classloader_caches` and can call that from our tests as needed.

It appears that what we were really relying on here is having the classloader cache cleared at interpreter shutdown; the `RuntimeTests` rely on this to avoid state pollution between tests. But we can do that easily enough in `Cinder_Fini`, so let's just move it there.

Reviewed By: DinoV

Differential Revision: D49397859

fbshipit-source-id: 0bc14bfb6c639b90110d7d3bbad699e2de2a3b0a