Reduce costs of checking if lazy imports are enabled
Summary: Checking if lazy imports is showing up as a significant (2%) cost in some workloads.  Those workloads are probably already less than ideal in that they're presumably doing imports in functions which is a bad idea.  But we can make that better by first checking to see if we're at a global scope which is super cheap.

Reviewed By: itamaro, Kronuz, cxxxs

Differential Revision: D49522414

fbshipit-source-id: 4ca46456309b36f4f75d03f493eaaf44068548e6