Fix option for disabling immortal instances
Summary:
There technically exists a build flag for disabling immortal instances in Cinder.  But this flag hasn't been maintained and has some brokenness around it.  There's also no way to enable/disable it.

This fixes up the issues with the flag and adds it to configure.

One thing I didn't address is fixing up various JIT tests - these got modified when this went in and we'd need some decent infrastructure to support both sets of results.

Reviewed By: carljm

Differential Revision: D49522323

fbshipit-source-id: 9a9e7906a8136240629fa514b161e9f82f614525